Add serialization to library
